#34 refactoring code
