+++ JobCareer/AppDelegate.swift
@@ -18,12 +18,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.rootViewController = appFlowController
+        window?.settingTheme()
         window?.makeKeyAndVisible()
         appFlowController.start()
 
         return true
     }
 }
+
+// MARK: - deleagte
 
 extension UIWindow: AppFlowControllerDelegate {
 
@@ -33,5 +36,12 @@ extension UIWindow: AppFlowControllerDelegate {
         }
         overrideUserInterfaceStyle = style
         AppDataHolder.colorTheme = value
+    }
+
+    func settingTheme() {
+        guard let style = UIUserInterfaceStyle(rawValue: AppDataHolder.colorTheme ?? 0) else {
+            return
+        }
+        overrideUserInterfaceStyle = style
     }
 }
